fix device busy bugs

+++ PyDockerMonitor/containerScheduler.py
@@ -128,7 +128,7 @@ class ContainerScheduler:
         self.hostToCommands[container.getHost()].append(containerCpuResponse)
         ##suspense memory incrementally
         while memory_value > 128:
-            memory_value = memory_value * 0.9
+            memory_value = int(memory_value * 0.9)
             memory_value_str = str(memory_value)+"m"
             cgroupMemoryKeyValue={
                                  "memory":{
